+++ src/math.hpp
@@ -11,7 +11,7 @@ namespace Math {
 float
 DegreesToRadians(float degrees)
 {
-    return degrees * M_PI / 180.0f;
+    return degrees * (float)M_PI / 180.0f;
 }
 
 } // namespace Math
